refactor: moved more components to screens
