+++ src/app/components/Popup/Popup.jsx
@@ -4,7 +4,7 @@ import { HashRouter, Switch, Route } from "react-router-dom";
 
 import utils from "../../../common/lib/utils";
 import Home from "../../screens/Home";
-import Unlock from "../Unlock";
+import Unlock from "../../screens/Unlock";
 import Loading from "../Loading";
 
 import "./styles.scss";
